+++ src/translations/en/purchase-patterns.php
@@ -11,4 +11,5 @@ return [
 	'There aren\'t any popular combinations yet!' => 'There aren\'t any popular combinations yet!',
 
 	'Purchased With' => 'Purchased With',
+	'Collating purchase pattern data' => 'Collating purchase pattern data',
 ];