+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-## [Unreleased] 1.1.0
+## 1.1.0 - 2019-06-17
 ### Added
+- Added "Extended" product query (requires fresh install)
+- Added job to collate purchase pattern data on plugin install
 - Added "Purchased With" section to product edit sidebar
 
 ## 1.0.0 - 2018-04-18